Fix assignment casts.
Assignment casts between different numeric types had gaps so that it was
not possible to assign a `float32` to an `int32`, etc. In general the
assignment casts logic seemed to rely on these casts being transitive
just as implicit casts.

The new assignment casts are non-transitive and have to be explicitly
defined for all valid pairs. However, in broad terms, implicit casts can
be used to connect the end points of an explicit assignment cast.